Begin work on pooled/intrusive lists.
The intention here is to allocate a large block of sequential pooled memory for combat log events (i.e. 20MB).
When an encounter begins, all combat events are logged out into the encounter log.
When the encounter completes, the combat log can be dumped out to a log file in the background.

This is mainly an experiment to build out systems similar to Details addon, Warcraft Logs for a game.